requirements.txt
alembic==1.11.1
black==23.3.0
blinker==1.6.2
click==8.1.3
dominate==2.7.0
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
python-dotenv==1.0.0
SQLAlchemy==2.0.15
tomli==2.0.1
typing_extensions==4.5.0
visitor==0.1.3
Werkzeug==2.3.4
